Supplemental data

These are several convenience datasets that make it easier to plot and analyze the anthromes data. They primarily include lookup tables to easily translate between numeric and character representations of anthromes, biomes, and other data, as well as color ramps for easy plotting. The biome data are derived from the 15-class potential natural vegetation dataset of Ramankutty, N. and Foley, J.A., 1999. Estimating historical changes in global land cover: Croplands from 1700 to 1992. Global biogeochemical cycles, 13(4), pp.997-1027, along with a simplification using only 9 classes.

anthrome_key

Format

An object of class tbl_df (inherits from tbl, data.frame) with 21 rows and 4 columns.

Value

Examples

time_steps_millennia

#>  [1] "10000BC" "9000BC"  "8000BC"  "7000BC"  "6000BC"  "5000BC"  "4000BC" 
#>  [8] "3000BC"  "2000BC"  "1000BC"  "0AD"     "1000AD"  "2000AD" 

time_steps_centuries

#>  [1] "0AD"    "100AD"  "200AD"  "300AD"  "400AD"  "500AD"  "600AD"  "700AD" 
#>  [9] "800AD"  "900AD"  "1000AD" "1100AD" "1200AD" "1300AD" "1400AD" "1500AD"
#> [17] "1600AD" "1700AD" "1800AD" "1900AD" "2000AD"

time_steps_years

#>  [1] "2000AD" "2001AD" "2002AD" "2003AD" "2004AD" "2005AD" "2006AD" "2007AD"
#>  [9] "2008AD" "2009AD" "2010AD" "2011AD" "2012AD" "2013AD" "2014AD" "2015AD"
#> [17] "2016AD" "2017AD"

time_steps

#>  [1] "10000BC" "9000BC"  "8000BC"  "7000BC"  "6000BC"  "5000BC"  "4000BC" 
#>  [8] "3000BC"  "2000BC"  "1000BC"  "0AD"     "100AD"   "200AD"   "300AD"  
#> [15] "400AD"   "500AD"   "600AD"   "700AD"   "800AD"   "900AD"   "1000AD" 
#> [22] "1100AD"  "1200AD"  "1300AD"  "1400AD"  "1500AD"  "1600AD"  "1700AD" 
#> [29] "1710AD"  "1720AD"  "1730AD"  "1740AD"  "1750AD"  "1760AD"  "1770AD" 
#> [36] "1780AD"  "1790AD"  "1800AD"  "1810AD"  "1820AD"  "1830AD"  "1840AD" 
#> [43] "1850AD"  "1860AD"  "1870AD"  "1880AD"  "1890AD"  "1900AD"  "1910AD" 
#> [50] "1920AD"  "1930AD"  "1940AD"  "1950AD"  "1960AD"  "1970AD"  "1980AD" 
#> [57] "1990AD"  "2000AD"  "2001AD"  "2002AD"  "2003AD"  "2004AD"  "2005AD" 
#> [64] "2006AD"  "2007AD"  "2008AD"  "2009AD"  "2010AD"  "2011AD"  "2012AD" 
#> [71] "2013AD"  "2014AD"  "2015AD"  "2016AD"  "2017AD"