Calculate anthrome summaries

A convenience function to create summaries of the total land area in each anthrome type at each time step, potentially subset by an additional variable.

anthrome_summary(anthromes, inputs, by = NULL)

hyde_summary(hyde, inputs, by = NULL)

Arguments

anthromes
A stars object containing the anthrome classification.

inputs
A stars object containing HYDE 3.2 fixed inputs.

by
a column in data by which to group the anthrome data before summary

hyde
A stars object containing HYDE 3.2 land use and population data.

Value

Functions

hyde_summary: Summarise HYDE 3.2 population time series.

Examples

# global summary
anthrome_summary(anthromes_med, inputs_med)

#> # A tibble: 19 x 7
#>    anthrome           `3000BC`  `2000BC`  `1000BC`     `0AD`  `1000AD`  `2000AD`
#>    <fct>                   [%]       [%]       [%]       [%]       [%]       [%]
#>  1 Urban              0.00000…  0.00000…  0.00000…  0.00000…  0.00000…  1.34354…
#>  2 Mixed settlements  0.01036…  0.01036…  0.01036…  0.02101…  0.12557…  1.19113…
#>  3 Rice villages      0.00000…  0.00000…  0.01068…  0.08541…  0.14963…  0.82291…
#>  4 Irrigated villag…  0.00000…  0.10593…  0.28775…  0.54654…  1.05498…  5.41930…
#>  5 Rainfed villages   0.03828…  0.04896…  0.07955…  0.71232…  0.14186…  7.15692…
#>  6 Pastoral villages  0.00000…  0.00000…  0.00000…  0.00000…  0.00000…  0.97033…
#>  7 Residential irri…  1.15440…  1.28037…  1.15855…  0.56932…  1.15223…  4.03996…
#>  8 Residential rain…  2.35864…  3.55433…  7.73114… 11.82416…  9.47797… 27.57045…
#>  9 Populated cropla…  4.98353…  5.68302…  5.20414…  4.61584…  5.16068… 10.57136…
#> 10 Remote croplands   0.52144…  1.46686…  2.47152…  3.03748…  3.70718…  2.08692…
#> 11 Residential rang…  0.00000…  0.07203…  0.05169…  1.08601…  0.38244…  2.43628…
#> 12 Populated rangel…  0.00000…  0.06130…  0.18712…  5.52944…  3.61036…  4.00107…
#> 13 Remote rangelands  0.23737…  0.28655…  1.82015…  1.32595…  2.42416…  3.98353…
#> 14 Residential wood…  0.04704…  0.32915…  1.17174…  2.01134…  2.73228…  1.57193…
#> 15 Populated woodla…  7.04758…  8.18503…  8.48347…  7.34081…  6.68319…  1.26049…
#> 16 Remote woodlands   4.18605…  2.13041…  0.51071…  0.45298…  0.25802…  0.10423…
#> 17 Inhabited drylan… 71.72798… 69.11729… 63.17898… 53.58786… 55.78579… 19.77959…
#> 18 Wild woodlands     0.12060…  0.11407…  0.10529…  0.09494…  0.09072…  0.06987…
#> 19 Wild drylands      7.56668…  7.55427…  7.53710…  7.15851…  7.06286…  5.62007…

# summary by regions
anthrome_summary(anthromes_med, inputs_med, by = regions)

#> # A tibble: 19 x 8
#>    regions  anthrome      `3000BC`  `2000BC` `1000BC`    `0AD` `1000AD` `2000AD`
#>    <fct>    <fct>              [%]       [%]      [%]      [%]      [%]      [%]
#>  1 Near Ea… Urban         0.00000…  0.00000…  0.0000…  0.0000…  0.0000…  1.3435…
#>  2 Near Ea… Mixed settl…  0.01036…  0.01036…  0.0103…  0.0210…  0.1255…  1.1911…
#>  3 Near Ea… Rice villag…  0.00000…  0.00000…  0.0106…  0.0854…  0.1496…  0.8229…
#>  4 Near Ea… Irrigated v…  0.00000…  0.10593…  0.2877…  0.5465…  1.0549…  5.4193…
#>  5 Near Ea… Rainfed vil…  0.03828…  0.04896…  0.0795…  0.7123…  0.1418…  7.1569…
#>  6 Near Ea… Pastoral vi…  0.00000…  0.00000…  0.0000…  0.0000…  0.0000…  0.9703…
#>  7 Near Ea… Residential…  1.15440…  1.28037…  1.1585…  0.5693…  1.1522…  4.0399…
#>  8 Near Ea… Residential…  2.35864…  3.55433…  7.7311… 11.8241…  9.4779… 27.5704…
#>  9 Near Ea… Populated c…  4.98353…  5.68302…  5.2041…  4.6158…  5.1606… 10.5713…
#> 10 Near Ea… Remote crop…  0.52144…  1.46686…  2.4715…  3.0374…  3.7071…  2.0869…
#> 11 Near Ea… Residential…  0.00000…  0.07203…  0.0516…  1.0860…  0.3824…  2.4362…
#> 12 Near Ea… Populated r…  0.00000…  0.06130…  0.1871…  5.5294…  3.6103…  4.0010…
#> 13 Near Ea… Remote rang…  0.23737…  0.28655…  1.8201…  1.3259…  2.4241…  3.9835…
#> 14 Near Ea… Residential…  0.04704…  0.32915…  1.1717…  2.0113…  2.7322…  1.5719…
#> 15 Near Ea… Populated w…  7.04758…  8.18503…  8.4834…  7.3408…  6.6831…  1.2604…
#> 16 Near Ea… Remote wood…  4.18605…  2.13041…  0.5107…  0.4529…  0.2580…  0.1042…
#> 17 Near Ea… Inhabited d… 71.72798… 69.11729… 63.1789… 53.5878… 55.7857… 19.7795…
#> 18 Near Ea… Wild woodla…  0.12060…  0.11407…  0.1052…  0.0949…  0.0907…  0.0698…
#> 19 Near Ea… Wild drylan…  7.56668…  7.55427…  7.5371…  7.1585…  7.0628…  5.6200…

# summarise hyde data
hyde_summary(hyde_med, inputs_med, by = regions)

#> # A tibble: 6 x 4
#>   regions   time          pop pop_percent
#>   <fct>     <fct>       <dbl>         [%]
#> 1 Near East 3000BC   2022683.    2.021059
#> 2 Near East 2000BC   3278506.    3.275873
#> 3 Near East 1000BC   4720362.    4.716572
#> 4 Near East 0AD      7327731.    7.321848
#> 5 Near East 1000AD   7670737.    7.664578
#> 6 Near East 2000AD 100080358.  100.000000